web_application/requirements.txt
# Веб-интерфейс
gradio~=5.49.1

# Инференс
onnxruntime~=1.23.1
numpy~=2.2.6
Pillow~=11.3.0